# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Call Calendar API
  version: 0.2.0
  description: |-
    Call Calendar — сервис «Запись на звонок».

    API-контракт (Design First): этот файл и соседние .tsp — единственный
    источник истины. Из них компилируется docs/api/openapi.yaml, по которому
    независимо реализуются бэкенд (Fastify) и фронтенд (React).

    Сценарии:
    - Гость: открывает /u/{slug}, смотрит типы событий и свободные слоты
    (в своей таймзоне на фронтенде), бронирует звонок.
    - Владелец: регистрируется, входит, управляет типами событий и доступностью,
    смотрит и отменяет предстоящие встречи.
...
paths:
  /api/availability:
    get:
      operationId: AvailabilityRoutes_getAvailability
      description: Текущие настройки доступности
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
      tags:
        - Owner
      security:
        - ApiKeyAuth: []
    put:
      operationId: AvailabilityRoutes_updateAvailability
      description: |-
        Полная замена настроек доступности.
        Интервалы каждого правила должны быть кратны 15 минутам и не пересекаться в рамках дня.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Availability'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Availability'
      security:
        - ApiKeyAuth: []
  /api/bookings:
    get:
      operationId: OwnerBookingRoutes_list
      description: Владелец смотрит список предстоящих встреч (только активные, от текущего момента)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
      tags:
        - Owner
      security:
        - ApiKeyAuth: []
  /api/bookings/{id}:
    delete:
      operationId: OwnerBookingRoutes_cancel
      description: Владелец отменяет встречу (перевод в статус cancelled, слот снова свободен)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Owner
      security:
        - ApiKeyAuth: []
  /api/event-types:
    get:
      operationId: OwnerEventTypeRoutes_list
      description: Список своих типов событий
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
      tags:
        - Owner
      security:
        - ApiKeyAuth: []
    post:
      operationId: OwnerEventTypeRoutes_create
      description: Создать тип события
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeWrite'
      security:
        - ApiKeyAuth: []
  /api/event-types/{id}:
    put:
      operationId: OwnerEventTypeRoutes_update
      description: Обновить тип события
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeWrite'
      security:
        - ApiKeyAuth: []
    delete:
      operationId: OwnerEventTypeRoutes_remove
      description: Удалить тип события (если нет активных будущих броней)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Owner
      security:
        - ApiKeyAuth: []
  /api/owners:
    post:
      operationId: OwnerRoutes_register
      description: |-
        Регистрация нового владельца календаря.
        При успехе устанавливается сессионная cookie (как при входе).
        409 — email или slug уже заняты.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '409':
          description: 409 — email или slug уже заняты при регистрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExistsError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Owners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCreate'
  /api/public/{slug}:
    get:
      operationId: PublicRoutes_getOwner
      description: Публичный профиль владельца для страницы /u/{slug}
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerPublic'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Guest
  /api/public/{slug}/bookings:
    post:
      operationId: PublicRoutes_createBooking
      description: |-
        Гость создаёт бронирование на свободный слот владельца.
        409 — слот занят (пересечение с другой бронью), в прошлом, вне расписания
        или дальше горизонта бронирования (14 дней).
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: '409 — конфликт: слот уже занят или находится в прошлом'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
  /api/public/{slug}/event-types:
    get:
      operationId: PublicRoutes_listEventTypes
      description: Типы событий этого владельца
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Guest
  /api/public/{slug}/slots:
    get:
      operationId: PublicRoutes_listSlots
      description: |-
        Свободные слоты за период [from, to] для выбранного типа события.
        Длина слота = durationMinutes типа. Горизонт — 14 дней включая сегодня.
        Даты from/to интерпретируются в таймзоне владельца.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: eventTypeId
          in: query
          required: true
          description: 'Тип события: определяет длительность слота'
          schema:
            type: integer
            format: int32
          explode: false
        - name: from
          in: query
          required: true
          description: Начало периода, YYYY-MM-DD
          schema:
            type: string
            format: date
          explode: false
        - name: to
          in: query
          required: true
          description: Конец периода, YYYY-MM-DD
          schema:
            type: string
            format: date
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '404':
          description: 404 — ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Guest
  /api/session:
    post:
      operationId: SessionRoutes_login
      description: Вход владельца. Устанавливает сессионную cookie.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '422':
          description: 422 — ошибка валидации входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreate'
    get:
      operationId: SessionRoutes_current
      description: Текущая сессия (проверка «я залогинен?»)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: 401 — нет активной сессии владельца или неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
      tags:
        - Session
      security:
        - ApiKeyAuth: []
    delete:
      operationId: SessionRoutes_logout
      description: Выход владельца. Удаляет сессию.
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Session
      security:
        - ApiKeyAuth: []
components:
  schemas:
    AlreadyExistsError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - already_exists
        message:
          type: string
      description: 409 — email или slug уже заняты при регистрации
    Availability:
      type: object
      required:
        - timezone
        - rules
      properties:
        timezone:
          type: string
          description: Часовой пояс владельца, IANA-идентификатор (например, Europe/Moscow)
        rules:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityRule'
      description: 'Настройки доступности владельца: часовой пояс + набор недельных правил'
      example:
        timezone: Europe/Moscow
        rules:
          - weekday: 1
            startTime: '10:00'
            endTime: '18:00'
          - weekday: 3
            startTime: '10:00'
            endTime: '14:00'
...
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - startsAt
        - endsAt
        - guestName
        - guestEmail
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int32
        eventTypeId:
          type: integer
          format: int32
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        guestName:
          type: string
        guestEmail:
          type: string
          format: email
        comment:
          type: string
          description: Комментарий гостя владельцу (тема звонка и т. п.)
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
      description: 'Бронирование: гость занял конкретный слот под конкретный тип события'
      example:
        id: 1
        eventTypeId: 1
        startsAt: '2026-07-27T10:00:00Z'
        endsAt: '2026-07-27T10:30:00Z'
        guestName: Иван Петров
        guestEmail: ivan@example.com
        comment: Хочу обсудить проект
        status: active
        createdAt: '2026-07-21T12:00:00Z'
    BookingCreate:
      type: object
      required:
        - eventTypeId
        - startsAt
        - guestName
        - guestEmail
      properties:
        eventTypeId:
          type: integer
          format: int32
        startsAt:
          type: string
          format: date-time
          description: Начало выбранного слота, UTC. Должно совпадать со слотом из GET …/slots.
        guestName:
          type: string
          minLength: 1
        guestEmail:
          type: string
          format: email
        comment:
          type: string
      description: Данные для создания бронирования гостем
    BookingStatus:
      type: string
      enum:
        - active
        - cancelled
    ConflictError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - slot_unavailable
        message:
          type: string
      description: '409 — конфликт: слот уже занят или находится в прошлом'
    EventType:
      type: object
      required:
        - id
        - name
        - durationMinutes
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 15
          maximum: 240
          description: |-
            Длительность встречи в минутах: 15–240, кратно 15.
            От неё зависят длина слота и шаг сетки при вычислении свободного времени.
      description: Тип события — то, на что гость записывается (например, «Вводный звонок»)
      example:
        id: 1
        name: Вводный звонок
        description: Знакомство и обсуждение задачи
        durationMinutes: 30
    EventTypeWrite:
      type: object
      required:
        - name
        - durationMinutes
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
          minimum: 15
          maximum: 240
      description: Создание / обновление типа события владельцем
...
    NotFoundError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - not_found
        message:
          type: string
      description: 404 — ресурс не найден
...
    OwnerCreate:
      type: object
      required:
        - name
        - email
        - password
        - slug
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        slug:
          type: string
          pattern: ^[a-z0-9]([a-z0-9-]{0,30}[a-z0-9])?$
      description: Данные регистрации нового владельца
    OwnerPublic:
      type: object
      required:
        - id
        - name
        - slug
        - timezone
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        slug:
          type: string
        timezone:
          type: string
          description: IANA-таймзона владельца (в ней заданы правила доступности)
      description: Публичный профиль владельца для гостевой страницы бронирования
      example:
        id: 1
        name: Кирилл Чистов
        slug: kirill
        timezone: Europe/Moscow
    Session:
      type: object
      required:
        - owner
      properties:
        owner:
          $ref: '#/components/schemas/Owner'
      description: Активная сессия
    SessionCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      description: Учётные данные владельца для входа
    Slot:
      type: object
      required:
        - startsAt
        - endsAt
      properties:
        startsAt:
          type: string
          format: date-time
          description: Начало слота, UTC
        endsAt:
          type: string
          format: date-time
          description: Конец слота, UTC (startsAt + durationMinutes типа события)
      description: |-
        Свободный слот. Вычисляется на лету из правил доступности и длительности
        выбранного типа события за вычетом активных бронирований — в базе не хранится.
      example:
        startsAt: '2026-07-27T10:00:00Z'
        endsAt: '2026-07-27T10:30:00Z'
    UnauthorizedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthorized
        message:
          type: string
      description: 401 — нет активной сессии владельца или неверные учётные данные
    ValidationError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - validation_error
        message:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
      description: 422 — ошибка валидации входных данных
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session
servers:
  - url: http://localhost:5001
    description: Локальный сервер разработки
    variables: {}